Refactor: Splie GameRunner
